backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
